Extract register id related types into register::id module;
